Improve speed and flexibility
Removed the template images dictionary and added an os.walk function. This will make it so we can remove and add images at will without changing the code, as long as the file ends in .jpg. Also removed some images to improve the speed, but kept enough images to provide accurate tracking.